requirements.txt
easygui==0.98.3
glcontext==2.3.6
glfw==2.5.3
imgui==1.4.1
moderngl==5.6.4
numpy==1.22.4
Pillow==9.1.1
PyOpenGL==3.1.0
